create return bike navigation graph
